Refactor cETH.ts and cECR20.ts into one file
