+++ subgraph.yaml
@@ -9,19 +9,19 @@ dataSources:
     network: mainnet
     source:
       address: "0x6c8c6b02e7b2be14d4fa6022dfd6d75921d90e4e"
-      abi: CErc20
+      abi: CToken
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.3
       language: wasm/assemblyscript
-      file: ./src/mappings/cERC20.ts
+      file: ./src/mappings/ctoken.ts
       entities:
         - User
         - Market
         - CTokenInfo
       abis:
-        - name: CErc20
-          file: ./abis/cERC20.json
+        - name: CToken
+          file: ./abis/ctoken.json
         - name: PriceOracle
           file: ./abis/priceOracle.json
         - name: PriceOracle2
@@ -46,19 +46,19 @@ dataSources:
     network: mainnet
     source:
       address: "0xf5dce57282a584d2746faf1593d3121fcac444dc"
-      abi: CErc20
+      abi: CToken
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.3
       language: wasm/assemblyscript
-      file: ./src/mappings/cERC20.ts
+      file: ./src/mappings/ctoken.ts
       entities:
         - User
         - Market
         - CTokenInfo
       abis:
-        - name: CErc20
-          file: ./abis/cERC20.json
+        - name: CToken
+          file: ./abis/ctoken.json
         - name: PriceOracle
           file: ./abis/priceOracle.json
         - name: PriceOracle2
@@ -83,19 +83,19 @@ dataSources:
     network: mainnet
     source:
       address: "0x4ddc2d193948926d02f9b1fe9e1daa0718270ed5"
-      abi: CEther
+      abi: CToken
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.3
       language: wasm/assemblyscript
-      file: ./src/mappings/cETH.ts
+      file: ./src/mappings/ctoken.ts
       entities:
         - User
         - Market
         - CTokenInfo
       abis:
-        - name: CEther
-          file: ./abis/cETH.json
+        - name: CToken
+          file: ./abis/ctoken.json
         - name: PriceOracle
           file: ./abis/priceOracle.json
         - name: PriceOracle2
@@ -118,19 +118,19 @@ dataSources:
     network: mainnet
     source:
       address: "0x158079ee67fce2f58472a96584a73c7ab9ac95c1"
-      abi: CErc20
+      abi: CToken
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.3
       language: wasm/assemblyscript
-      file: ./src/mappings/cERC20.ts
+      file: ./src/mappings/ctoken.ts
       entities:
         - User
         - Market
         - CTokenInfo
       abis:
-        - name: CErc20
-          file: ./abis/cERC20.json
+        - name: CToken
+          file: ./abis/ctoken.json
         - name: PriceOracle
           file: ./abis/priceOracle.json
         - name: PriceOracle2
@@ -155,19 +155,19 @@ dataSources:
     network: mainnet
     source:
       address: "0x39aa39c021dfbae8fac545936693ac917d5e7563"
-      abi: CErc20
+      abi: CToken
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.3
       language: wasm/assemblyscript
-      file: ./src/mappings/cERC20.ts
+      file: ./src/mappings/ctoken.ts
       entities:
         - User
         - Market
         - CTokenInfo
       abis:
-        - name: CErc20
-          file: ./abis/cERC20.json
+        - name: CToken
+          file: ./abis/ctoken.json
         - name: PriceOracle
           file: ./abis/priceOracle.json
         - name: PriceOracle2
@@ -192,19 +192,19 @@ dataSources:
     network: mainnet
     source:
       address: "0xc11b1268c1a384e55c48c2391d8d480264a3a7f4"
-      abi: CErc20
+      abi: CToken
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.3
       language: wasm/assemblyscript
-      file: ./src/mappings/cERC20.ts
+      file: ./src/mappings/ctoken.ts
       entities:
         - User
         - Market
         - CTokenInfo
       abis:
-        - name: CErc20
-          file: ./abis/cERC20.json
+        - name: CToken
+          file: ./abis/ctoken.json
         - name: PriceOracle
           file: ./abis/priceOracle.json
         - name: PriceOracle2
@@ -229,19 +229,19 @@ dataSources:
     network: mainnet
     source:
       address: "0xb3319f5d18bc0d84dd1b4825dcde5d5f7266d407"
-      abi: CErc20
+      abi: CToken
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.3
       language: wasm/assemblyscript
-      file: ./src/mappings/cERC20.ts
+      file: ./src/mappings/ctoken.ts
       entities:
         - User
         - Market
         - CTokenInfo
       abis:
-        - name: CErc20
-          file: ./abis/cERC20.json
+        - name: CToken
+          file: ./abis/ctoken.json
         - name: PriceOracle
           file: ./abis/priceOracle.json
         - name: PriceOracle2
